refactor: single-source toUtcIsoDate into @mulmoclaude/common (#2480)
