+++ docs/shared-utils.md
@@ -16,7 +16,8 @@ This catalog only covers **cross-cutting** helpers — formatters, error helpers
 | `server/utils/time.ts`     | `SUBPROCESS_PROBE_TIMEOUT_MS`, `SUBPROCESS_WORK_TIMEOUT_MS`, `CLI_SUBPROCESS_TIMEOUT_MS`, `STARTUP_FAILURE_FORCE_EXIT_MS`, `DEV_PLUGIN_WATCH_DEBOUNCE_MS` | Named timeout presets for subprocess / CLI / startup paths. Add a new constant rather than passing a literal. |
 | `src/utils/format/date.ts` | `formatDate`, `formatDateTime`, `formatTime`, `formatShortDate`, `formatShortTime`, `formatMonthYear`, `formatSmartTime`, `formatRelativeTime`            | User-facing date display in Vue Views. Prefer over inline `toLocaleString()`.                                 |
 | `src/utils/format/date.ts` | `isSameDay`, `isToday`                                                                                                                                    | Day-boundary comparisons.                                                                                     |
-| `server/utils/date.ts`     | `toUtcIsoDate(...)` and friends                                                                                                                            | Server-side UTC date normalisation (`YYYY-MM-DD`). Distinct from the frontend display formatters above.        |
+| `@mulmoclaude/common`      | `toUtcIsoDate(timestamp)`                                                                                                                                  | **The one implementation** (#2480) of `Date` → `YYYY-MM-DD` in UTC — isomorphic leaf, so any tier imports the same copy (host via the `server/utils/date.ts` re-export, x-plugin directly). The `@receptron/task-scheduler` copy stays local on purpose — see the duplicates table. |
+| `server/utils/date.ts`     | `toLocalIsoDate`, `toUtcIsoDate` (re-export), `isoDateOnly`, `isValidIsoDate`                                                                              | Server-side date normalisation (`YYYY-MM-DD`). Distinct from the frontend display formatters above.            |
 | `packages/core/src/google/collectionDateTime.ts` (`@mulmoclaude/core/google`) | `toCollectionDateTime(value)` | RFC3339-with-zone / date-only → the `YYYY-MM-DDTHH:MM[:SS]` shape a collection `datetime` field is linted and rendered against (#2310): the zone designator is dropped (wall clock kept, never converted to host-local) and an all-day date gains `T00:00`. Pure. Used by the Google Calendar → collection sync; reuse it for any other importer writing into a `datetime` field instead of re-deriving the stripping. |
 
 ## Errors
@@ -218,7 +219,7 @@ rule requiring it.
 | `slugify` | 3 host/frontend + 1 canonical plugin | **Resolved for the plugins (#2405)** | The three plugin copies (chart / html / mulmoscript) are gone — they now delegate to `slugifyArtifact` in `@mulmoclaude/core/artifacts`, which adopts chart's strip → cap → strip order (the old html/mulmoscript cap-then-strip differed only in length for 120+char titles starting with punctuation; pinned by a boundary test). The host `server/utils/slug.ts` (URL slugs, crypto-hash non-ASCII fallback) and the frontend memory/topic slugs stay separate by design. |
 | `escapeHtml` | 4 | no | Two spellings (switch-map vs. chained `.replace`), identical output. Canonical: `@mulmoclaude/core/wiki`'s `render.ts` — the host `wikiEmbeds.ts` copy is gone, it and its consumers import the core one (#2460). Remaining: spotify-plugin, markdown-utils `mermaidExtension`, an inline arrow in `server/api/routes/runtime-plugin.ts`. |
 | `yearMonthUtc` | 1 | **Resolved (#2460)** | Canonical: `@mulmoclaude/core/artifacts` (browser-safe, #2405). The host copy in `server/utils/files/naming.ts` is gone — `naming.ts` and the image / attachment / spreadsheet stores import the core one. |
-| `toUtcIsoDate` | 3 | no | `server/utils/date.ts`, `packages/scheduler/src/date.ts`, `x-plugin/src/internal.ts`. |
+| `toUtcIsoDate` | 1 canonical + `@receptron/task-scheduler` | no | Canonical: `@mulmoclaude/common` (#2480). Host `server/utils/date.ts` re-exports it; the `x-plugin/src/internal.ts` copy is gone — x-plugin imports common directly (its zero-dep KEEP rationale ended when #2467 added the `@mulmoclaude/common` dep). The `packages/scheduler/src/date.ts` copy is **deliberate**: `@receptron/task-scheduler` is an independently published, dependency-free leaf package, and taking a common dep for one 6-line pure function would cost more than the duplication. |
 | `safeResponseText` | 2 | no | `server/utils/http.ts` and `x-plugin/src/internal.ts`. |
 | `asyncHandler` | 2 | no | `server/utils/asyncHandler.ts` and `accounting-plugin/src/server/http.ts`. |
 | Time constants | 13+ | no | Values agree everywhere. `server/workspace/wiki-pages/snapshot.ts` re-declares `ONE_DAY_MS` inside `server/` (**accidental** — `server/utils/time.ts` is one import away). The frontend has no catalogued source; `src/composables/useRunElapsed.ts` and `src/utils/agent/formatElapsed.ts` each declare their own. |
